Specify staging profile ID
Need to do it at the module level to simplify CI system. If the profile
ID is specified in the CI system it will apply to all modules and it
only needs to apply to the API. The implementation has a separate
staging profile ID. There are ways around this in the CI system but this
should be simpler.

Signed-off-by: Mark Thomas <markt@apache.org>